requirements.txt
evdev==1.9.2
fire==0.7.1
pillow==12.1.0
pynput==1.8.1
python-xlib==0.33
six==1.17.0
termcolor==3.3.0
